Added player selection and play round function
The play round function simulates a round of rock, paper, or scissors.
The function compares two arguments that are passed, the player
selection and the computer selection, the comparisons are the same as
any regular rock, paper, or scissors game(ie. rock beats scissors,
paper beats rock etc.). If the player wins a message is sent through
the browser console, another message is sent if the player loses or if
the player ties with the computer.