Moved JS into dedicated file and updated game
RPS Game doesn't run for 5 turns and can run as many times as the player
would like, it also now has a ui where the player can visually see the
options to choose. The result of the round will now show directly under
the options(will adjust to a more appropriate location).